renamed to pagingallocator
